Tests for strategies in separate files
