+++ runtime/test/planner-tests.js
@@ -16,16 +16,9 @@ import {assert} from './chai-web.js';
 import Manifest from '../manifest.js';
 import Recipe from '../recipe/recipe.js';
 import ConvertConstraintsToConnections from '../strategies/convert-constraints-to-connections.js';
-import InitPopulation from '../strategies/init-population.js';
 import MapSlots from '../strategies/map-slots.js';
 import ResolveRecipe from '../strategies/resolve-recipe.js';
-import MatchParticleByVerb from '../strategies/match-particle-by-verb.js';
 import MatchRecipeByVerb from '../strategies/match-recipe-by-verb.js';
-import SearchTokensToParticles from '../strategies/search-tokens-to-particles.js';
-import GroupHandleConnections from '../strategies/group-handle-connections.js';
-import CombinedStrategy from '../strategies/combined-strategy.js';
-import CreateDescriptionHandle from '../strategies/create-description-handle.js';
-import FallbackFate from '../strategies/fallback-fate.js';
 import MessageChannel from '../message-channel.js';
 import InnerPec from '../inner-PEC.js';
 import Particle from '../particle.js';
@@ -203,326 +196,6 @@ describe('Planner', function() {
     // the thread index instead.
     assert.equal(plans[0].groupIndex, 0);
     assert.equal(plans[1].groupIndex, 1);
-  });
-});
-
-import InitSearch from '../strategies/init-search.js';
-describe('InitSearch', async () => {
-  it('initializes the search recipe', async () => {
-    let arc = new Arc({id: 'test-plan-arc', context: {}});
-    arc._search = 'search';
-    let initSearch = new InitSearch(arc);
-    let inputParams = {generated: [], generation: 0};
-    let results = await initSearch.generate(inputParams);
-    assert.equal(results.length, 1);
-    assert.equal(results[0].score, 0);
-  });
-});
-
-describe('InitPopulation', async () => {
-  it('penalizes resolution of particles that already exist in the arc', async () => {
-    let manifest = await Manifest.parse(`
-      schema Product
-
-      particle A in 'A.js'
-        A(in Product product)
-
-      recipe
-        create as v1
-        A
-          product <- v1`);
-    let recipe = manifest.recipes[0];
-    assert(recipe.normalize());
-    let arc = new Arc({id: 'test-plan-arc', context: {recipes: [recipe]}});
-    await arc.instantiate(recipe);
-    let ip = new InitPopulation(arc);
-
-    let inputParams = {generated: [], generation: 0};
-    let results = await ip.generate(inputParams);
-    assert.equal(results.length, 1);
-    assert.equal(results[0].score, 0);
-  });
-});
-
-describe('ConvertConstraintsToConnections', async () => {
-
-  it('fills out an empty constraint', async () => {
-    let recipe = (await Manifest.parse(`
-      schema S
-      particle A
-        A(inout S b)
-      particle C
-        C(inout S d)
-
-      recipe
-        A.b -> C.d`)).recipes[0];
-    let inputParams = {generated: [{result: recipe, score: 1}]};
-    let cctc = new ConvertConstraintsToConnections({pec: {}});
-    let results = await cctc.generate(inputParams);
-    assert.equal(1, results.length);
-    let {result, score} = results[0];
-    assert.deepEqual(result.toString(),
-`recipe
-  create as view0 // S
-  A as particle0
-    b = view0
-  C as particle1
-    d = view0`);
-  });
-
-  it('does not cause an input only handle to be created', async () => {
-    let recipe = (await Manifest.parse(`
-      schema S
-      particle A
-        A(in S b)
-      particle C
-        C(in S d)
-
-      recipe
-        A.b -> C.d`)).recipes[0];
-    let inputParams = {generated: [{result: recipe, score: 1}]};
-    let cctc = new ConvertConstraintsToConnections({pec: {}});
-    let results = await cctc.generate(inputParams);
-    assert.equal(0, results.length);
-  });
-
-  it('can resolve input only handle connection with a mapped handle', async () => {
-    let recipe = (await Manifest.parse(`
-      schema S
-      particle A
-        A(in S b)
-      particle C
-        C(in S d)
-
-      recipe
-        map as v0
-        A.b -> C.d`)).recipes[0];
-    let inputParams = {generated: [{result: recipe, score: 1}]};
-    let cctc = new ConvertConstraintsToConnections({pec: {}});
-    let results = await cctc.generate(inputParams);
-    assert.equal(1, results.length);
-  });
-
-  it('can create handle for input and output handle', async () => {
-    let createRecipe = async (constraint1, constraint2) => (await Manifest.parse(`
-      schema S
-      particle A
-        A(in S b)
-      particle C
-        C(in S d)
-      particle E
-        E(out S f)
-
-      recipe
-        ${constraint1}
-        ${constraint2}`)).recipes[0];
-    let verify = async (constraint1, constraint2) => {
-      let recipe = await createRecipe(constraint1, constraint2);
-      let inputParams = {generated: [{result: recipe, score: 1}]};
-      let cctc = new ConvertConstraintsToConnections({pec: {}});
-      let results = await cctc.generate(inputParams);
-      assert.equal(1, results.length, `Failed to resolve ${constraint1} & ${constraint2}`);
-    };
-    // Test for all possible combination of connection constraints with 3 particles.
-    let constraints = [['A.b -> C.d', 'C.d -> A.b'], ['A.b -> E.f', 'E.f -> A.b'], ['C.d -> E.f', 'E.f -> C.d']];
-    for (let i = 0; i < constraints.length; ++i) {
-      for (let j = 0; j < constraints.length; ++j) {
-        if (i == j) continue;
-        for (let ii = 0; ii <= 1; ++ii) {
-          for (let jj = 0; jj <= 1; ++jj) {
-            await verify(constraints[i][ii], constraints[j][jj]);
-          }
-        }
-      }
-    }
-  });
-
-  it('fills out a constraint, reusing a single particle', async () => {
-    let recipe = (await Manifest.parse(`
-      schema S
-      particle A
-        A(inout S b)
-      particle C
-        C(inout S d)
-
-      recipe
-        A.b -> C.d
-        C`)).recipes[0];
-    let inputParams = {generated: [{result: recipe, score: 1}]};
-    let cctc = new ConvertConstraintsToConnections({pec: {}});
-    let results = await cctc.generate(inputParams);
-    assert.equal(1, results.length);
-    let {result, score} = results[0];
-    assert.deepEqual(result.toString(),
-`recipe
-  create as view0 // S
-  A as particle0
-    b = view0
-  C as particle1
-    d = view0`);
-  });
-
-  it('fills out a constraint, reusing a single particle (2)', async () => {
-    let recipe = (await Manifest.parse(`
-      schema S
-      particle A
-        A(inout S b)
-      particle C
-        C(inout S d)
-
-      recipe
-        A.b -> C.d
-        A`)).recipes[0];
-    let inputParams = {generated: [{result: recipe, score: 1}]};
-    let cctc = new ConvertConstraintsToConnections({pec: {}});
-    let results = await cctc.generate(inputParams);
-    assert.equal(1, results.length);
-    let {result, score} = results[0];
-    assert.deepEqual(result.toString(),
-`recipe
-  create as view0 // S
-  A as particle0
-    b = view0
-  C as particle1
-    d = view0`);
-  });
-
-
-  it('fills out a constraint, reusing two particles', async () => {
-    let recipe = (await Manifest.parse(`
-      schema S
-      particle A
-        A(inout S b)
-      particle C
-        C(inout S d)
-
-      recipe
-        A.b -> C.d
-        C
-        A`)).recipes[0];
-    let inputParams = {generated: [{result: recipe, score: 1}]};
-    let cctc = new ConvertConstraintsToConnections({pec: {}});
-    let results = await cctc.generate(inputParams);
-    assert.equal(1, results.length);
-    let {result, score} = results[0];
-    assert.deepEqual(result.toString(),
-`recipe
-  create as view0 // S
-  A as particle0
-    b = view0
-  C as particle1
-    d = view0`);
-  });
-
-  it('fills out a constraint, reusing two particles and a view', async () => {
-    let recipe = (await Manifest.parse(`
-      schema S
-      particle A
-        A(inout S b)
-      particle C
-        C(inout S d)
-
-      recipe
-        A.b -> C.d
-        use as v1
-        C
-          d = v1
-        A`)).recipes[0];
-    let inputParams = {generated: [{result: recipe, score: 1}]};
-    let cctc = new ConvertConstraintsToConnections({pec: {}});
-    let results = await cctc.generate(inputParams);
-    assert.equal(1, results.length);
-    let {result, score} = results[0];
-    assert.deepEqual(result.toString(),
-`recipe
-  use as view0 // S
-  A as particle0
-    b = view0
-  C as particle1
-    d = view0`);
-  });
-
-  it('fills out a constraint, reusing two particles and a view (2)', async () => {
-    let recipe = (await Manifest.parse(`
-      schema S
-      particle A
-        A(inout S b)
-      particle C
-        C(inout S d)
-
-      recipe
-        A.b -> C.d
-        use as v1
-        C
-        A
-          b = v1`)).recipes[0];
-    let inputParams = {generated: [{result: recipe, score: 1}]};
-    let cctc = new ConvertConstraintsToConnections({pec: {}});
-    let results = await cctc.generate(inputParams);
-    assert.equal(1, results.length);
-    let {result, score} = results[0];
-    assert.deepEqual(result.toString(),
-`recipe
-  use as view0 // S
-  A as particle0
-    b = view0
-  C as particle1
-    d = view0`);
-  });
-
-  it('removes an already fulfilled constraint', async () => {
-    let recipe = (await Manifest.parse(`
-      schema S
-      particle A
-        A(inout S b)
-      particle C
-        C(inout S d)
-
-      recipe
-        A.b -> C.d
-        use as v1
-        C
-          d = v1
-        A
-          b = v1`)).recipes[0];
-    let inputParams = {generated: [{result: recipe, score: 1}]};
-    let cctc = new ConvertConstraintsToConnections({pec: {}});
-    let results = await cctc.generate(inputParams);
-    assert.equal(1, results.length);
-    let {result, score} = results[0];
-    assert.deepEqual(result.toString(), `recipe
-  use as view0 // S
-  A as particle0
-    b = view0
-  C as particle1
-    d = view0`);
-  });
-
-  it('verifies affordance', async () => {
-    let recipes = (await Manifest.parse(`
-      schema S
-      particle A in 'A.js'
-        A(out S b)
-        affordance voice
-        consume root
-      particle C in 'C.js'
-        C(in S d)
-        affordance voice
-        consume root
-      particle E in 'E.js'
-        E(in S f)
-        consume root
-
-      recipe
-        A.b -> C.d
-      recipe
-        A.b -> E.f
-    `)).recipes;
-    let inputParams = {generated: [{result: recipes[0], score: 1}, {result: recipes[1], score: 1}]};
-    let cctc = new ConvertConstraintsToConnections({pec: {slotComposer: {affordance: 'voice'}}});
-    let results = await cctc.generate(inputParams);
-    assert.equal(results.length, 1);
-    assert.deepEqual(results[0].result.particles.map(p => p.name), ['A', 'C']);
   });
 });
 
@@ -784,290 +457,6 @@ describe('AssignOrCopyRemoteViews', function() {
   });
 });
 
-describe('SearchTokensToParticles', function() {
-  it('particles by verb strategy', async () => {
-    let manifest = (await Manifest.parse(`
-      particle SimpleJumper in 'A.js'
-        jump()
-      particle StarJumper in 'AA.js'
-        jump()
-      particle GalaxyFlyer in 'AA.js'
-        fly()
-      particle Rester in 'AA.js'
-        rest()
-
-      recipe
-        search \`jump or fly or run and Rester\`
-    `));
-    let arc = StrategyTestHelper.createTestArc('test-plan-arc', manifest, 'dom');
-    let recipe = manifest.recipes[0];
-    assert(recipe.normalize());
-    assert(!recipe.isResolved());
-    let inputParams = {generated: [], terminal: [{result: recipe, score: 1}]};
-    let stp = new SearchTokensToParticles(arc);
-    let results = await stp.generate(inputParams);
-    assert.equal(results.length, 2);
-    assert.deepEqual([['GalaxyFlyer', 'Rester', 'SimpleJumper'],
-                      ['GalaxyFlyer', 'Rester', 'StarJumper']], results.map(r => r.result.particles.map(p => p.name).sort()));
-    assert.deepEqual(['fly', 'jump', 'rester'], results[0].result.search.resolvedTokens);
-    assert.deepEqual(['and', 'or', 'or', 'run'], results[0].result.search.unresolvedTokens);
-  });
-});
-
-describe('MatchRecipeByVerb', function() {
-  it('removes a particle and adds a recipe', async () => {
-    let manifest = await Manifest.parse(`
-      recipe
-        particle can jump
-
-      schema Feet
-      schema Energy
-
-      particle JumpingBoots in 'A.js'
-        JumpingBoots(in Feet f, in Energy e)
-      particle FootFactory in 'B.js'
-        FootFactory(out Feet f)
-      particle NuclearReactor in 'C.js'
-        NuclearReactor(out Energy e)
-
-      recipe jump
-        JumpingBoots.f <- FootFactory.f
-        JumpingBoots.e <- NuclearReactor.e
-    `);
-
-    let arc = StrategyTestHelper.createTestArc('test-plan-arc', manifest, 'dom');
-    let inputParams = {generated: [{result: manifest.recipes[0], score: 1}]};
-    let mrv = new MatchRecipeByVerb(arc);
-    let results = await mrv.generate(inputParams);
-    assert.equal(results.length, 1);
-    assert.equal(results[0].result.particles.length, 0);
-    assert.deepEqual(results[0].result.toString(), 'recipe\n  JumpingBoots.e -> NuclearReactor.e\n  JumpingBoots.f -> FootFactory.f');
-  });
-  it('plays nicely with constraints', async () => {
-    let manifest = await Manifest.parse(`
-      schema S
-      particle P in 'A.js'
-        P(out S p)
-      particle Q in 'B.js'
-        Q(in S q)
-
-      recipe
-        P.p -> Q.q
-        particle can a
-
-      recipe a
-        P
-    `);
-
-    let arc = StrategyTestHelper.createTestArc('test-plan-arc', manifest, 'dom');
-    let inputParams = {generated: [{result: manifest.recipes[0], score: 1}]};
-    let mrv = new MatchRecipeByVerb(arc);
-    let results = await mrv.generate(inputParams);
-    assert.equal(results.length, 1);
-    let cctc = new ConvertConstraintsToConnections(arc);
-    results = await cctc.generate({generated: results});
-    assert.equal(results.length, 1);
-    assert.deepEqual(results[0].result.toString(),
-`recipe
-  create as view0 // S
-  P as particle0
-    p -> view0
-  Q as particle1
-    q <- view0`);
-  });
-});
-
-describe('MatchParticleByVerb', function() {
-  let manifestStr = `
-    schema Energy
-    schema Height
-    particle SimpleJumper in 'A.js'
-      jump(in Energy e, out Height h)
-      affordance dom
-      consume root
-    particle StarJumper in 'AA.js'
-      jump(in Energy e, inout Height h)
-      affordance dom
-      consume root
-    particle VoiceStarJumper in 'AA.js'  // wrong affordance
-      jump(in Energy e, out Height h)
-      affordance voice
-      consume root
-    particle GalaxyJumper in 'AA.js'  // wrong connections
-      jump(in Energy e)
-      affordance dom
-      consume root
-    particle StarFlyer in 'AA.js'  // wrong verb
-      fly()
-
-    recipe
-      use as height
-      use as energy
-      particle can jump
-        * = height
-        * <- energy
-  `;
-
-  it('particles by verb strategy', async () => {
-    let manifest = (await Manifest.parse(manifestStr));
-    let arc = StrategyTestHelper.createTestArc('test-plan-arc', manifest, 'dom');
-    // Apply MatchParticleByVerb strategy.
-    let inputParams = {generated: [{result: manifest.recipes[0], score: 1}]};
-    let mpv = new MatchParticleByVerb(arc);
-    let results = await mpv.generate(inputParams);
-    assert.equal(results.length, 3);
-    // Note: view connections are not resolved yet.
-    assert.deepEqual(['GalaxyJumper', 'SimpleJumper', 'StarJumper'], results.map(r => r.result.particles[0].name).sort());
-  });
-
-  it('particles by verb recipe fully resolved', async () => {
-    let manifest = (await Manifest.parse(manifestStr));
-    let recipe = manifest.recipes[0];
-    recipe.handles[0].mapToView({id: 'test1', type: manifest.findSchemaByName('Height').entityClass().type});
-    recipe.handles[1].mapToView({id: 'test2', type: manifest.findSchemaByName('Energy').entityClass().type});
-
-    let arc = StrategyTestHelper.createTestArc('test-plan-arc', manifest, 'dom');
-
-    // Apply all strategies to resolve recipe where particles are referenced by verbs.
-    let planner = new Planner();
-    planner.init(arc);
-    let plans = await planner.plan(1000);
-
-    assert.equal(2, plans.length);
-    assert.deepEqual([['SimpleJumper'], ['StarJumper']],
-                     plans.map(plan => plan.particles.map(particle => particle.name)));
-  });
-
-  describe('GroupHandleConnections', function() {
-    let schemaAndParticlesStr = `
-schema Thing
-schema OtherThing
-particle A
-  A(in Thing ithingA1)
-particle B
-  B(in Thing ithingB1, in Thing ithingB2, in [OtherThing] iotherthingB1)
-particle C
-  C(in Thing ithingC1, out Thing othingC2, inout [OtherThing] iootherthingC1)
-particle D
-  D(in Thing ithingD1, in Thing ithingD2, out Thing othingD3)
-    `;
-    it('group in and out view connections', async () => {
-      // TODO: add another Type view connections to the recipe!
-      let manifest = (await Manifest.parse(`
-${schemaAndParticlesStr}
-recipe
-  A
-  B
-  C
-  D
-      `));
-      let inputParams = {generated: [{result: manifest.recipes[0], score: 1}]};
-      let arc = StrategyTestHelper.createTestArc('test-plan-arc', manifest, 'dom');
-      arc._search = 'ShowCollection and chooser alsoon recommend';
-      let ghc = new GroupHandleConnections(arc);
-
-      let results = await ghc.generate(inputParams);
-      assert.equal(results.length, 1);
-      let recipe = results[0].result;
-      assert.equal(4, recipe.handles.length);
-      // Verify all connections are bound to handles.
-      assert.isUndefined(recipe.handleConnections.find(hc => !hc.handle));
-      // Verify all handles have non-empty connections list.
-      assert.isUndefined(recipe.handles.find(v => v.connections.length == 0));
-    });
-  });
-  describe('CombinedStrategy', function() {
-    it('combined strategy with search tokens and group handle connections', async () => {
-      let manifest = (await Manifest.parse(`
-        schema Energy
-        schema Height
-        particle Energizer in 'A.js'
-          prepare(out Energy energy)
-        particle Jumper in 'AA.js'
-          jump(in Energy energy, out Height height)
-
-        recipe
-          search \`prepare and jump\`
-      `));
-      manifest.recipes[0].normalize();
-      let inputParams = {generated: [{result: manifest.recipes[0], score: 1}], terminal: []};
-      let arc = StrategyTestHelper.createTestArc('test-plan-arc', manifest, 'dom');
-      let strategy = new CombinedStrategy([
-        new SearchTokensToParticles(arc),
-        new GroupHandleConnections(arc),
-      ]);
-
-      let results = await strategy.generate(inputParams);
-      assert.equal(results.length, 1);
-      let recipe = results[0].result;
-      assert.equal(2, recipe.particles.length);
-      assert.equal(1, recipe.handles.length);
-      assert.equal(2, recipe.handles[0].connections.length);
-    });
-  });
-});
-describe('FallbackFate', function() {
-  it('fallback for search based recipe', async () => {
-    let manifest = (await Manifest.parse(`
-      schema Thing
-      particle DoSomething in 'AA.js'
-        DoSomething(in Thing inthing, out Thing outthing)
-
-      recipe
-        search \`DoSomething DoSomethingElse\`
-        use as view0
-        use as view1
-        DoSomething as particle0
-          inthing <- view0
-          outthing -> view1
-    `));
-    let recipe = manifest.recipes[0];
-    recipe.handles.forEach(v => v._originalFate = '?');
-    assert(recipe.normalize());
-    let arc = StrategyTestHelper.createTestArc('test-plan-arc', manifest, 'dom');
-    let inputParams = {generated: [{result: manifest.recipes[0], score: 1}], terminal: []};
-    let strategy = new FallbackFate(arc);
-
-    // no resolved search tokens.
-    let results = await strategy.generate(inputParams);
-    assert.equal(results.length, 0);
-
-    // Resolved a search token and rerun strategy.
-    recipe.search.resolveToken('DoSomething');
-    results = (await strategy.generate(inputParams));
-    assert.equal(results.length, 1);
-    let plan = results[0].result;
-    assert.equal(plan.handles.length, 2);
-    assert.equal('map', plan.handles[0].fate);
-    assert.equal('copy', plan.handles[1].fate);
-  });
-
-  it('ignore non-search unresolved recipe', async () => {
-    // Same as above, but the recipe doesn't originate from user search query.
-    let manifest = (await Manifest.parse(`
-      schema Thing
-      particle DoSomething in 'AA.js'
-        DoSomething(in Thing inthing, out Thing outthing)
-
-      recipe
-        use as view0
-        use as view1
-        DoSomething as particle0
-          inthing <- view0
-          outthing -> view1
-    `));
-    let recipe = manifest.recipes[0];
-    recipe.handles.forEach(v => v._originalFate = '?');
-    assert(recipe.normalize());
-    let arc = StrategyTestHelper.createTestArc('test-plan-arc', manifest, 'dom');
-    let inputParams = {generated: [{result: manifest.recipes[0], score: 1}], terminal: []};
-
-    let strategy = new FallbackFate(arc);
-    let results = await strategy.generate(inputParams);
-    assert.equal(results.length, 0);
-  });
-});
-
 describe('Type variable resolution', function() {
   let loadAndPlan = async (manifestStr) => {
     let loader = {
@@ -1263,29 +652,6 @@ view MyThings1 of [Thing1] #mythings1 in 'things.json'
 schema Thing2
 view MyThings2 of [Thing2] #mythings2 in 'things.json'`);
   */
-  });
-});
-
-describe('CreateDescriptionHandle', function() {
-  it('descriptions handle created', async () => {
-    let manifest = (await Manifest.parse(`
-      schema Description
-      particle DoSomething in 'AA.js'
-        DoSomething(out [Description] descriptions)
-
-      recipe
-        DoSomething as particle0
-    `));
-    let recipe = manifest.recipes[0];
-    let inputParams = {generated: [{result: manifest.recipes[0], score: 1}], terminal: []};
-    let strategy = new CreateDescriptionHandle();
-    let results = (await strategy.generate(inputParams));
-
-    assert.equal(results.length, 1);
-    let plan = results[0].result;
-    assert.equal(plan.handles.length, 1);
-    assert.equal('create', plan.handles[0].fate);
-    assert.isTrue(plan.isResolved());
   });
 });
